pyproject.toml
[tool.poetry]
name = "prism-api"
version = "0.1.0"
description = "AI-driven knowledge management platform"
authors = ["hwuiwon <hwuiwon.kim@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.1"
llama-index = "^0.7.16"
python-dotenv = "^1.0.0"
boto3 = "^1.28.16"
uvicorn = "^0.23.2"
mergepythonclient = "^0.2.1"
unstructured = {extras = ["csv", "doc", "docx", "html", "pdf", "ppt", "pptx", "rtf", "txt", "xlsx"], version = "^0.9.0"}
ray = {extras = ["default"], version = "^2.6.2"}
langchain = {extras = ["llms"], version = "^0.0.251"}


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
pre-commit = "^3.3.3"
pylint = "^2.17.5"
ruff = "^0.0.281"
isort = "^5.12.0"

[tool.ruff]
line-length = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
